Rename directory and files that contain config values
