+++ packages/beacon-node/test/unit-mainnet/network/gossip/scoringParameters.test.ts
@@ -2,7 +2,7 @@ import {expect} from "chai";
 import {TopicScoreParams} from "@chainsafe/libp2p-gossipsub/score";
 import {ATTESTATION_SUBNET_COUNT, ForkName, SLOTS_PER_EPOCH} from "@lodestar/params";
 import {createBeaconConfig} from "@lodestar/config";
-import {mainnetChainConfig} from "@lodestar/config/presets";
+import {mainnetChainConfig} from "@lodestar/config/configs";
 import {computeGossipPeerScoreParams, gossipScoreThresholds} from "../../../../src/network/gossip/scoringParameters.js";
 import {stringifyGossipTopic} from "../../../../src/network/gossip/topic.js";
 import {GossipType} from "../../../../src/network/index.js";
