+++ packages/config/src/chainConfig/configs/mainnet.ts
