+++ packages/config/src/chainConfig/configs/minimal.ts
