+++ packages/config/src/chainConfig/networks/chiado.ts
@@ -2,7 +2,7 @@
 import {fromHexString as b} from "@chainsafe/ssz";
 import {PresetName} from "@lodestar/params";
 import {ChainConfig} from "../types.js";
-import {chainConfig as mainnet} from "../presets/mainnet.js";
+import {chainConfig as mainnet} from "../configs/mainnet.js";
 
 // Chiado beacon chain config:
 // https://github.com/gnosischain/configs/blob/main/chiado/config.yaml
