+++ packages/config/src/chainConfig/networks/ropsten.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import {fromHexString as b} from "@chainsafe/ssz";
 import {ChainConfig} from "../types.js";
-import {chainConfig as mainnet} from "../presets/mainnet.js";
+import {chainConfig as mainnet} from "../configs/mainnet.js";
 
 // Ropsten beacon chain config:
 // https://github.com/eth-clients/merge-testnets/blob/main/ropsten-beacon-chain/config.yaml
