+++ packages/config/src/configs.ts
@@ -1,0 +1,4 @@
+import {chainConfig as mainnetChainConfig} from "./chainConfig/configs/mainnet.js";
+import {chainConfig as minimalChainConfig} from "./chainConfig/configs/minimal.js";
+
+export {mainnetChainConfig, minimalChainConfig};
