--- packages/config/src/presets.ts
@@ -1,4 +1,0 @@
-import {chainConfig as mainnetChainConfig} from "./chainConfig/presets/mainnet.js";
-import {chainConfig as minimalChainConfig} from "./chainConfig/presets/minimal.js";
-
-export {mainnetChainConfig, minimalChainConfig};
